Add refactored analysis code
This code can be used to run all analyses presented in:

Chapman, et al., “Reachability Analysis as a Design Tool for Stormwater Systems,” in 6th IEEE Conference on Technologies for Sustainability, November 2018.

This code has been refactored from the original version to improve documentation and provide ease of scenario selection, with the hopeful expectation that others may benefit from its publication.

All results have been tested to be equivalent to the results presented in the paper using Matlab 2018a.

Co-authored-by: Margaret Chapman <chapmanm@berkeley.edu>
Co-authored-by: Victoria Cheng <chengv@berkeley.edu>